course refactor and message added

+++ src/main/java/com/lifecourse/course_service/modules/course/infrastructure/persistence/CourseRepositoryAdapter.java
@@ -2,6 +2,9 @@ package com.lifecourse.course_service.modules.course.infrastructure.persistence;
 
 import com.lifecourse.course_service.modules.course.web.dto.CourseResponse;
 import com.lifecourse.course_service.modules.course.web.dto.CreateCourseRequest;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
@@ -14,8 +17,7 @@ import java.util.concurrent.atomic.AtomicReference;
 @Component
 public class CourseRepositoryAdapter {
     private final CourseRepository courseRepository;
-    AtomicBoolean isSaved = new AtomicBoolean(false);
-    AtomicReference<CourseEntity> courseEntityAtomicReference = new AtomicReference<CourseEntity>();
+    @Autowired
     public CourseRepositoryAdapter(CourseRepository courseRepository) {
         this.courseRepository = courseRepository;
     }
@@ -31,13 +33,8 @@ public class CourseRepositoryAdapter {
     }
 
 
-    public CourseEntity findById(Long id) {
-        Optional<CourseEntity> resultEntity=courseRepository.findById(id);
-         if(resultEntity.isPresent()){
-             return resultEntity.get();
-         }
-         return null;
-
+    public Optional<CourseEntity> findById(Long id) {
+       return  courseRepository.findById(id);
     }
 
     public CourseEntity save(CourseEntity course) {
@@ -45,7 +42,10 @@ public class CourseRepositoryAdapter {
 
     }
 
-    public void delete(Long id) {
+    public void delete(Long id) throws EntityNotFoundException {
+        if (!courseRepository.existsById(id)) {
+            throw new EntityNotFoundException("Course with id " + id + " not found");
+        }
         courseRepository.deleteById(id);
     }
     private CourseResponse mapToCourseResponse(CourseEntity course) {
@@ -61,19 +61,16 @@ public class CourseRepositoryAdapter {
     }
 
     public CourseResponse createCourse(CreateCourseRequest request) {
-        CourseEntity course = CourseEntity.builder()
-                .title(request.title())
-                .description(request.description())
-                .categories(request.category())
-                .level(request.level())
-                .price(request.price())
-                .instructorId(request.instructorId())
-                .durationMinutes(request.durationMinutes())
-                .published(Boolean.TRUE.equals(request.published()))
-                .build();
+        CourseEntity course = new CourseEntity()
+                .setTitle(request.title())
+                .setDescription(request.description())
+                .setCategories(request.category())
+                .setLevel(request.level())
+                .setPrice(request.price())
+                .setDurationMinutes(request.durationMinutes())
+                .setPublished(Boolean.TRUE.equals(request.published()));
+        return  mapToCourseResponse(courseRepository.save(course));
 
-        CourseEntity savedCourse = courseRepository.save(course);
-        return mapToCourseResponse(savedCourse);
 
     }
 
@@ -85,7 +82,6 @@ public class CourseRepositoryAdapter {
                             .setCategories(request.category())
                             .setLevel(request.level())
                             .setPrice(request.price())
-                            .setInstructorId(request.instructorId())
                             .setDurationMinutes(request.durationMinutes())
                             .setPublished(Boolean.TRUE.equals(request.published()));
